Move some frontend utility functions to __utils__
Otherwise tests from the frontend are executed 3 times.
