+++ __tests__/__utils__/index.ts
@@ -23,3 +23,12 @@
 export * from './expect-helper'
 export * from './kv'
 export * from './services'
+
+export enum Backend {
+  Frontend = 'frontend',
+  Legacy = 'legacy',
+}
+
+export function setupProbabilityFor(backend: Backend) {
+  global.FRONTEND_PROBABILITY = backend === Backend.Frontend ? '1' : '0'
+}
